+++ HISTORY.md
@@ -8,6 +8,8 @@ https://github.com/josdejong/jsoneditor
 - Implemented a navigation bar showing the path. Thanks @meirotstein.
 - Implemented a status bar showing cursor location.
   Thanks @meirotstein.
+- Implemented repairing JSON objects containing left and right single
+  and double quotes (which you get when typing a JSON object in Word).
 
 
 ## 2017-09-16, version 5.9.6